% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircularBoxplots-package.R
\docType{package}
\name{CircularBoxplots-package}
\alias{CircularBoxplots}
\alias{CircularBoxplots-package}
\title{CircularBoxplots: Grouped Circular Boxplots for Circular Data}
\description{
Create grouped boxplots for comparison of circular data across groups.
}
\author{
\strong{Maintainer}: Josh Berlinski \email{josh.berlinski@gmail.com} (Fan Dai <fand@mtu.edu> [aut])

}
\keyword{internal}
